requirements.txt
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
langchain-core>=0.1.40
langchain>=0.1.14
langgraph>=0.0.30
langchain_openai>=0.1.1
langchainhub>=0.1.15
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0